+++ aera/src/commands.rs
@@ -3,7 +3,7 @@ pub enum Command {
     // Absolute move
     MovJ(i64, i64, i64, i64),
     // Relative move
-    Move(i64, i64, i64, i64),
+    Move(f64, f64, f64, f64),
     // Grab with the robot
     Grab,
     // Release what the robot is holding
